backend1/requirements.txt
# Core Python Libraries
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==1.4.46
psycopg2-binary==2.9.6
alembic==1.10.3

# Authentication and Email
python-dotenv==1.0.0
email-validator==2.0.0
secure-smtpd==1.1.1

# OCR and PDF Processing
pdf2image
paddleocr==2.6.1.3
opencv-python-headless==4.7.0.72
numpy==1.24.3

# Machine Learning and NLP
llama-index==0.7.12
huggingface-hub==0.14.1
transformers==4.29.2
torch==2.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Additional Utilities
pydantic==1.10.7
jsonschema==4.17.3
requests==2.30.0